redis/requirements.txt
redis
fastapi
uvicorn
python-dotenv